// src/types/domain.types.ts
export enum UserRole {
  PASSENGER = 'PASSENGER',
  SUPPORT_EXEC = 'SUPPORT_EXEC',
  ADMIN = 'ADMIN',
  SUPER_ADMIN = 'SUPER_ADMIN',
}
// ...
export enum CouponDiscountType {
  PERCENTAGE = 'PERCENTAGE',
  FLAT = 'FLAT',
}

export enum CmsContentType {
  HOMEPAGE = 'homepage',
  FAQ = 'faq',
  LEGAL_PRIVACY = 'legal_privacy',
  LEGAL_TERMS = 'legal_terms',
  SETTINGS = 'settings',
}